net-misc/dhcp: pull upstream ebuild 4.2.2-r1
This version is only used in certain networking tests, and not in
production images. It is needed for linux-3.0 as it has rewritten
build system which doesn't break on linux3.

It needs to be explicitly unmasked by package.unmask, because our
version of portage snapshot hard masks it.

Compared to upstream, this ebuild is x86, arm, amd64 keyworded.

BUG=chromium-os:18442
TEST=build it, trybot chromiumos-sdk

Change-Id: I746ceac1cce924c28e55810b83a04e9833f65c8b
Reviewed-on: http://gerrit.chromium.org/gerrit/6806
Tested-by: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>