sys-shells/bash: import from upstream Gentoo
The only reason we have bash in the cros overlay is so that we have it
not install a /bin/sh symlink pointing to bash in the board root.  Once
we move that handling to chromeos-base, we can drop all the bash ebuild
customizations.

BUG=chromium-os:21901
TEST=`emerge chromeos-base bash` and see /bin/sh -> bash
TEST=`emerge-amd64-generic chromeos-base dash` and see /bin/sh -> dash
TEST=`emerge-arm-generic chromeos-base dash` and see /bin/sh -> dash
TEST=`emerge-x86-generic chromeos-base dash` and see /bin/sh -> dash

Change-Id: I1df4a75ac314ed308817b4cdf61564ac0174ad63
Reviewed-on: https://gerrit.chromium.org/gerrit/11409
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>