cmake: upgrade the cmake Portage package
The CL upgrades dev-util/cmake to version 2.8.4-r1 on amd64. The
update is necessary because I need to compile a library (OpenCV)
for a factory test. Besides that, the current version (2.6.4) is
more than 2 years old and therefore worthwhile for an upgrade.

A corresponding CL on the keywords file will be carried out before
this one.

BUG=chromium-os:23747
TEST='emerge cmake' still works

Change-Id: I29154af89d373297d692f70c77a35cbdd2da5737
Reviewed-on: https://gerrit.chromium.org/gerrit/12484
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>