sshfs-fuse: pull in version 2.4 from upstream
The upstream version 2.4 adds `-o slave`, which allows us to spawn
sshfs attached to the server end of an SSH pipe. As a result files
on your workstation can easily be pushed to a chromium-os device,
rather than needing to open a terminal on the device and SSH back.

BUG=chromium-os:25146
TEST=clean build and install test image on x86-alex
TEST=ran chrome over sshfs with and without -o slave

Change-Id: I0d4641bb4a4b0310ceb0afcb5042c7873c05357f
Reviewed-on: https://gerrit.chromium.org/gerrit/17724
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>