smdk-dltool: Produce correct return code in event of error
At present if a download fails, we don't report this in the return code.
This means there is no indication of failure to cros_bundle_firmware.

Correct the return code in this case, with a new patch.

BUG=chrome-os-partner:9633
TEST=manual:
$ cros_bundle_firmware -b daisy -w usb --bl1 <fname>
where fname is an old BL1 which breaks USB download.
See that the failure to download BL2 is now reported and
cros_bundle_firmware stops with an error.

Change-Id: I3bb966bf3c1a9009cfba4064da872a8dd8cb6043
Reviewed-on: https://gerrit.chromium.org/gerrit/23145
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>