coreutils: upgraded package to upstream (and drop realpath)
Upgraded sys-apps/coreutils to version 8.20 on amd64, arm, and x86

This version includes the realpath package, so drop that from the
tree so people don't try to install/depend on it anymore.

This also includes a fix for the issue that caused it to be reverted
before -- broken `cp --no-preserve=mode`.

BUG=chromium-os:24204
TEST=`cbuildbot chromiumos-sdk` worked
TEST=`cbuildbot {amd64,arm,x86}-generic-full` worked

Change-Id: I473dee144dddacefe4ad60279497ce7e8e8c11c7
Previous-Reviewed-on: https://gerrit.chromium.org/gerrit/39312
Reviewed-on: https://gerrit.chromium.org/gerrit/39967
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Harring <ferringb@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>