add(app-emulation/nova-agent): initial commit
This adds the nova agent for Rackspace Cloud and its dependencies. Quite
a bit of work went into getting xen-tools to build properly as it has a
quirky ld version script which needed patching and also the tests are
not working and had to be disabled.

The two patches I wrote are:
	app-emulation/xen-tools/files/disable-tests.patch
	app-emulation/xen-tools/files/fix-map-file-ld-errors.patch

Luckily xen-tools is only needed to build pyxenstore, otherwise it is a
huge mess of stuff to keep on disk.