fix(eclass/cros-kernel/x86_64_defconfig) shrink the kernel configuration
This removes a bunch of hardware-specific kernel configuration options
that are not needed in a virtual machine.  It also removes reiserfs
support as well.

Compressed kernel image has shrunk by 290Kb, and we have way fewer
kernel modules being built, and loaded, as well (what in the world was
loading the agp drivers at boot?!)