fix(sys-kernel/coreos-kernel) no more firmware files
The kernel package should not be building any firmware files, as they
are not needed in virtual machines, _and_ the firmware file locations
are not versioned, preventing multiple kernels to be installed at the
same time.

This change adds a patch to the kernel to not traverse into the firmware
directory at all, and patches the eclass to not do the firmare_install
build process.