fix(coreos): Get version from manifest, export COREOS_ names
The version will now be tracked in the manifest so the old
coreos_version.sh needs to source it from there. Also export variables
with a COREOS_ prefix in addition to CHROMEOS_ since the latter name
isn't really correct. This is particularly true for the data published
via lsb_release.