feat(profiles): Add systemd profile from my Gentoo systemd-only overlay.
No need for me to maintain a similar profile in two entirely different
ways. This is also one tiny step towards cleaning up our profiles in
general. Original here: https://github.com/marineam/systemd-only-overlay

As part of this change the baselayout dependency on openrc is now
handled via a use flag instead of package.provided. We didn't previously
include a virtual/init package but Gentoo has one and I needed it for my
generic systemd-only overlay so might as well include it here if it is
needed in the future.