churn(profiles): Move arch/amd64
The only CoreOS specific things in this profile appear to have been
a few use flags for the SDK. Switch to the upstream amd64 profile after
moving these flags around in the appropriate target profiles.

The one function change here is that the delta_generator use flag is now
properly disabled outside of the SDK as it was in chromeos.