add(sys-firmware/edk2-ovmf): Add UEFI firmware for QEMU
To use either run qemu with:
  -bios efi-bios.bin
Or:
  -L path/to/usr/share/edk2-ovmf

The latter is useful if you want to use the OVMF built by the SDK
outside of the SDK. Using UEFI with the QEMU instance inside the CoreOS
SDK is kinda awkward because serial doesn't work and UEFI uses a VGA
graphics framebuffer mode, not VGA text mode so -curses doesn't work
either. In any case, it is a start.