fix(sys-apps/baselayout): Use custom script to generate /etc/group
I misunderstood the documentation for systemd-tmpfiles, if a string to
write to a file is provided it will always write it, even if the file
already exists and doesn't need to be created. This means that using
tmpfiles to initialize /etc/group results appending each boot.

Instead use a little script instead, also initialize passwd and shadow
so that the `passwd` command works for the core user.

Make use of the systemd eclass where applicable.