add(app-emulation/xenserver-pv-version): Fake XenServer PV drivers
XenServer expects virtual machines to run their own version of the Xen
para-virtualized drivers and will not attach disks otherwise. This is
pretty silly considering the ones bundled in Linux are perfectly fine.
In order to work around this write the latest version of XenServer to
special variables in xenstore. This will be included in all images
rather than OEM since this could impact any XenServer user and I want to
write to xenstore as early as possible, possibly even in the initrd.