bump(sys-kernel/coreos-kernel): Update to v3.14.5, update panic settings
In addition to bumping to the latest stable this tweaks the default
settings for panics and lockups so that they do not need to be set in
sysctl configs. We needed to update these settings anyway and setting it
here means they will apply before boot gets around loading sysctl confs.

Previously the sysctl options set the hardlockup threshold to 5 seconds
and the softlockup threshold to 10 seconds. This is perhaps a bit too
aggressive, just use the default values of 10 and 20 seconds.

Also sysctl enabled an immediate reboot after panic, using a friendlier
60 seconds gives users a chance to see the error before the reboot.