profiles: disable calling sync(2) after installing packages
This is just a safety feature for real Gentoo systems on the off chance
that they crash moments after a package is installed. There isn't any
such concern for the CoreOS SDK so just disable it. A normal build_image
is slightly faster with this but I suspect it may help the build host
which is usually running many emerge jobs in parallel.