coreos-kernel: enable support for common special input devices
One user had the misfortune of having two keyboards, both of which have
special HID drivers but neither of which we enable. (Apple and Logitech)

The generic HID driver ignores devices with special drivers even if the
drivers are not built. This patch enables all of the likely keyboard
related drivers that are normally selected by CONFIG_EXPERT. Hopefully
that covers everything that is usually expected to "Just Work".