baselayout: Add OEM lib directories to default search path
Although this shouldn't strictly be required, OEM binaries can use RPATH
or RUNPATH, there are plenty of corner cases that waste time. For
example if a library in OEM links to another OEM the linker verification
performed by build_packages won't catch it since RPATH would be provided
by the binary that happens to use those libraries, not the library files
themselves. Such tediousness isn't worth wasting more time over and as
long as OEM is after the normal system libraries this should be safe.