coreos-kernel: fix up build and source symlinks
The symlinks the kernel build adds to the modules directory are always
absolute, forever preserving the value of $ROOT a binary package was
built with. Instead use a relative link for source, pointing to the
location coreos-sources gets installed to. For build just create a
directory containing .config which the linux-info eclass can use to
check kernel features during builds, allowing us to migrate away from
having to set up the /var/cache build directory even when the kernel
wasn't built.

For both cases they are irrelevant for production builds and can be
confusing to users so mask them.

This eclass now assumes it is used in conjunction with USE=symlink-usr
and installs modules directly to /usr/lib/modules instead of depending
on the /lib symlink. This makes no significant impact other than
silencing a warning portage emits about the relative source symlink.