coreos-kernel: make build ccache friendly across versions
The absolute path to the source tree includes the kernel's version
number and that path then lands in the kernel source code via the
numerous __FILE__ macros in the kernel source. So if the kernel source
version changes, even with a minor revision only, the entire cache can
be invalidated.

Reshuffle the build tree so that $S is a directory of symlinks to the
source and a nested build directory, ensuring the entire build can use
relative paths. A directory of symlinks instead of a single directory
symlink is required to prevent make from resolving the source directory
path to the original versioned path.