glibc: apply CoreOS tweaks
This time around our changes take a slightly different approach. Instead
of editing the ebuild to undo things done in the common eblit code the
eblits are edited directly, allowing for better customization. Also,
managing locales is now different. Inside the SDK all locales are built
at compile time, avoiding the need for locale-gen which is slow and
complicates setup of the SDK chroot. For cross-compiled targets (CoreOS
itself) final images don't include locale info either, again no need for
locale-gen.

The SDK is a bit bigger now, the compressed tarball is 20MB bigger and
200MB bigger extracted on disk. Considering how big it already is this
shouldn't be that big of a deal.