coreos-kernel: convert to a two stage kernel build
The new coreos-modules ebuild does the bulk of the work, building and
installing all modules to /usr/lib/modules/... and is now the home of
the kernel config. This in turn allows the coreos-kernel ebuild to
depend on coreos-firmware and allowing dracut to handle kernel modules
and firmware instead of appending them to the cpio ourselves.