coreos-base/oem-ec2-compat: use ttyS0 for console
On HVM instances, tty0 is actually a graphical interface. Since the
interface exposed is non-interactive, it's not possible to scroll
through the logs, which results in lots of missing information. Setting
the console to ttyS0, on the other hand, sends the output to the system
log.

PV instances don't have this issue since they are xen-based and use
hvc0.