app-emulation/google-compute-engine: use a config template
The configuration template needs to be written by the service
itself in the current setup.  The host's /etc is mounted into the
ACI to provide all of the user's system configuration and to allow
the GCE scripts to make their modifications, so the template can't
be included in the ACI.  It can't be written to the host's /etc
either, since it is read-only in the SDK when OEM packages are
being installed.