sys-auth/sssd: CoreOSify the service
There is no need to fork the daemon with systemd, so run in the
foreground for Type=simple.  This also writes logs to stderr by
default, so this service no longer needs a special /var/log path.
The syslog.target unit does not exist either, but SSSD can
can optionally query NSCD, so order it after nscd.service.