sys-apps/systemd: upgrade to v234
This syncs most of the ebuild with Gentoo's, except it omits the
meson build part for now.

A side effect of this is that lxml will be built as a dependency,
but it is not used since systemd now requires Python 3.  This will
be fixed with meson support, since meson requires Python 3 as well.