app-backup/casync: add initial ebuild
This is a fairly boring meson ebuild. I had to apply a patch to make it
build.

This is included in the sdk for the torcx packaging work, where we'd
like to get a digest of a given torcx pkg. `casync digest` works well
for that and can be pointed at a directory easily enough.