eclass/coreos-cargo: Restructure how base libs are built
This changes the rustc wrapper script to actually use cargo's
RUSTC_WRAPPER variable, and the logic in the script now uses the
--target option to detect when it is cross-compiling.  If it is, it
transparently adds the --sysroot option to the rustc command so the
target libraries are used in the board as its sysroot.  This allows
dropping any distinction between host-compatible targets.

In doing this, the cross-compiler wrapper scripts had to be moved
from the rustlib package to the eclass for all builds, since no
longer having a --sysroot on every call would resurrect failures
due to LD_LIBRARY_PATH.

The rustlib package itself is now also building proc_macro, using
its target output directory in the library search path to find the
bootstrapping libraries as it is building them.