dev-db/etcdctl: Bump 3.3.1 to 3.3.2
This drops the Go version specification since it never worked.  It
needs to be defined before importing the eclass to have an effect.
Pinning the build to upstream's tested Go verion is likely not as
important for the client program as the server, which is run in a
container anyway.