sys-libs/glibc: apply CoreOS changes
The structure of the ebuild changed a bit but our changes are pretty
much the same as before.

We keep our gshadow patch in 2.25 since the directory is still used by
Gentoo.

This also moves the C UTF-8 locale patch to 2.26 since it hasn't changed
as of Fedora 27 (its upstream).