app-torcx/docker: Omit Docker package revisions from dependencies
Docker's related packages are pinned to the specific versions that
upstream bundles with each release.  Since our revision bumps are
typically only for backporting patches or changing configuration in
those same versions, we don't need to explicitly depend on their
revisions.  This still leaves the revision dependencies for ltdl
and tini since we handle those a bit differently, and this will
require special attention for package changes.