profiles: Enable libidn2 in systemd
Since glibc 2.29 wants to use the system's libidn2 for resolving,
we might as well be consistent with systemd.

Note that glibc can work without it if libidn2 is not installed, so
this change effectively enables it for both glibc and systemd.  The
glibc dependency on libidn2 from upstream had to be dropped to keep
the bootstrapping process working.