sys-apps/systemd: Adapt the sed command to new version
After systemd version 243, the lines that sed is removing now include
an exclamation mark, this commit adapts the sed command to accept both
formats.

These lines need to be removed so that the configuration set in
/usr/lib/tmpfiles.d/pam.conf file can be used (duplicate entries will
lead to systemd not doing anything).

The pam.conf configuration is shipped by of the pam package:
https://github.com/flatcar-linux/coreos-overlay/blob/flatcar-master/sys-libs/pam/files/tmpfiles.d/pam.conf

Fixes: flatcar-linux/Flatcar#75